requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.13.1,<3.14.0
django-allauth>=0.50.0,<0.51.0
django-rest-auth>=0.9.5,<0.10.0
flake8>=4.0.1,<4.1.0
psycopg2>=2.9.3,<2.10.0
Pillow>=9.0,<9.1.0
